% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betanbb.R
\name{BETANBB}
\alias{BETANBB}
\title{Beta-Negative Binomial Bayesian Dynamic Model}
\usage{
BETANBB(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
Conjugate Bayesian dynamic model for count time series with a negative
binomial observation distribution and a Beta prior on the success probability
parameter. This extends the conjugate updating framework of Harvey &
Fernandes (1989) to the Beta-Negative Binomial family. The Beta prior is
updated at each time step using a discount factor \code{w}. Forecasts are available
as samples simulating from the model forward in time.
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)

fc_ts <- ts |>
  model(BETANBB(value)) |>
  forecast(h = "7 days")

fc_ts |> print()


if (requireNamespace("ggtime", quietly = TRUE)) {
  library(ggtime)
  fc_ts |> autoplot(ts)
}

}
\references{
Harvey, A. C., & Fernandes, C. (1989). Time series models for count or
qualitative observations. \emph{Journal of Business & Economic Statistics},
7(4), 407--417. \doi{10.1080/07350015.1989.10509750}.
}
