% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_tweedie.R
\name{dtweedie}
\alias{dtweedie}
\title{Compute Tweedie Density}
\usage{
dtweedie(x, mean = 1, dispersion = 1, power = 1.5, log = FALSE)
}
\arguments{
\item{x}{The values the density is evaluated at.}

\item{mean}{Mean parameter of the distribution.}

\item{dispersion}{Dispersion parameter of the distribution.}

\item{power}{Power parameter of the distribution.}

\item{log}{Whether to return the logarithm.}
}
\description{
Evaluates the Tweedie density via the series expansion of Dunn & Smyth
(2005), dispatching to the C++ implementation \code{tweedieDensity}.
}
\keyword{internal}
