% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.intermittent-package.R
\docType{package}
\name{fable.intermittent-package}
\alias{fable.intermittent}
\alias{fable.intermittent-package}
\title{fable.intermittent: Forecasting Models for Intermittent Time Series}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extends the 'fable' framework to support forecasting methods specifically designed for intermittent time series data, where demand occurs sporadically with many zero values. All methods produce probabilistic forecasts returned as 'distributional' objects. The returned forecasts can be used to evaluate accuracy, plot and print the results seamlessly with 'fable'. The methods include: Harvey, Fernandes (1989) \doi{10.1080/07350015.1989.10509750}, Willemain, Smart, Schwarz (2004) \doi{10.1016/S0169-2070(03)00013-X}, Zhou, Viswanathan (2011) \doi{10.1016/j.ijpe.2010.09.021}, Snyder, Ord, Beaumont (2012) \doi{10.1016/j.ijforecast.2011.03.009}, Kolassa (2016) \doi{10.1016/j.ijforecast.2015.12.004}, Hasni, Aguir, Babai, Jemai (2019) \doi{10.1080/00207543.2018.1424375}, Damato, Azzimonti, Corani (2025) \doi{10.1016/j.ijforecast.2025.10.001}, Sbrana (2025) \doi{10.1080/01605682.2025.2569661}, Sbrana, Babai (2026) \doi{10.1016/j.ejor.2026.06.009}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StefanoDamato/fable.intermittent}
  \item Report bugs at \url{https://github.com/StefanoDamato/fable.intermittent/issues}
}

}
\author{
\strong{Maintainer}: Stefano Damato \email{stefanodamato128@gmail.com} (\href{https://orcid.org/0009-0001-0225-3736}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Stefano Damato \email{stefanodamato128@gmail.com} (\href{https://orcid.org/0009-0001-0225-3736}{ORCID}) [copyright holder]
  \item Lorenzo Zambon \email{lorenzo.zambon@idsia.ch} (\href{https://orcid.org/0000-0002-8939-993X}{ORCID})
  \item Dario Azzimonti \email{dario.azzimonti@gmail.com} (\href{https://orcid.org/0000-0001-5080-3061}{ORCID})
}

}
\keyword{internal}
