% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betanbb.R
\name{fitted.BETANBB}
\alias{fitted.BETANBB}
\title{Extract fitted values from a BETANBB model}
\usage{
\method{fitted}{BETANBB}(object, ...)
}
\arguments{
\item{object}{A model for which fitted values are required.}

\item{...}{Not used.}
}
\value{
A numeric vector of fitted values.
}
\description{
Extract fitted values from a BETANBB model
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)
fit <- model(ts, BETANBB(value))
fitted(fit)
}
