% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betanbb.R
\name{forecast.BETANBB}
\alias{forecast.BETANBB}
\title{Forecast a BETANBB model}
\usage{
\method{forecast}{BETANBB}(object, new_data, specials = NULL, times = 10000, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing future index values to forecast.}

\item{specials}{Passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}.}

\item{times}{The number of sample paths to use in estimating the forecast
distribution.}

\item{...}{Other arguments passed to methods}
}
\value{
A distribution vector of class \code{dist_sample}.
}
\description{
Produces forecast distributions from a fitted BETANBB model using
simulation.
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)
fit <- model(ts, BETANBB(value))
forecast(fit, h = "7 days")

}
