% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empdistr.R
\name{forecast.EMPDISTR}
\alias{forecast.EMPDISTR}
\title{Forecast an EMPDISTR model}
\usage{
\method{forecast}{EMPDISTR}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing future index values to forecast.}

\item{specials}{Passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}.}

\item{...}{Other arguments passed to methods}
}
\value{
A distribution vector of class \code{dist_sample}.
}
\description{
Produces forecast distributions by repeating the empirical distribution
estimated from the training data at each forecast horizon.
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)
fit <- model(ts, EMPDISTR(value))
forecast(fit, h = "7 days")
}
