% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsd.R
\name{forecast.PARAMSD}
\alias{forecast.PARAMSD}
\title{Forecast a PARAMSD model}
\usage{
\method{forecast}{PARAMSD}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing future index values to forecast.}

\item{specials}{Passed by \code{\link[fabletools:forecast.mdl_df]{fabletools::forecast.mdl_df()}}.}

\item{...}{Other arguments passed to methods}
}
\value{
A distribution vector. The class depends on the static distribution
fitted by the \code{PARAMSD} method.
}
\description{
Produces forecast distributions from a fitted PARAMSD model.
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)
fit <- model(ts, PARAMSD(value))
forecast(fit, h = "7 days")

}
