% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empsd.R
\name{generate.EMPSD}
\alias{generate.EMPSD}
\title{Generate sample paths from an EMPSD model}
\usage{
\method{generate}{EMPSD}(x, new_data, specials = NULL, ...)
}
\arguments{
\item{x}{A fitted \code{EMPSD} model object.}

\item{new_data}{A tsibble containing future index values to forecast.}

\item{specials}{Passed by \code{\link[fabletools:forecast.mdl_df]{fabletools::forecast.mdl_df()}}.}

\item{...}{Other arguments passed to methods}
}
\value{
A \code{new_data} tibble with a \code{.sim} column of simulated values.
}
\description{
Generate sample paths from an EMPSD model
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)
fit <- model(ts, EMPSD(value))
generate(fit, h = 7)
}
