% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnarma.R
\name{generate.NNARMA}
\alias{generate.NNARMA}
\title{Generate sample paths from a NNARMA model}
\usage{
\method{generate}{NNARMA}(x, new_data, specials = NULL, ...)
}
\arguments{
\item{x}{A fitted \code{NNARMA} model object.}

\item{new_data}{A tsibble containing future index values to forecast.}

\item{specials}{Passed by \code{\link[fabletools:forecast.mdl_df]{fabletools::forecast.mdl_df()}}.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of future paths from a dataset using a fitted model.
}
\description{
Generate sample paths from a NNARMA model
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)
fit <- model(ts, NNARMA(value))
generate(fit, new_data = tsibble::new_data(ts, 7))
}
