% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twees_occ_damped.R
\name{generate.TWEES_OCC_DAMPED}
\alias{generate.TWEES_OCC_DAMPED}
\title{Generate sample paths from a TWEES model}
\usage{
\method{generate}{TWEES_OCC_DAMPED}(x, new_data, specials = NULL, ...)
}
\arguments{
\item{x}{A fitted \code{TWEES} model object.}

\item{new_data}{A tsibble containing future index values to forecast.}

\item{specials}{Passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of future paths from a dataset using a fitted model.
}
\description{
Generate sample paths from a TWEES model
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)
fit <- model(ts, TWEES(value))
generate(fit, new_data = tsibble::new_data(ts, 7))
}
