% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twees_occ_damped_p0_fixed.R
\name{generate.TWEES_OCC_DAMPED_p0_FIXED}
\alias{generate.TWEES_OCC_DAMPED_p0_FIXED}
\title{Generate sample paths from a TWEES_OCC_DAMPED_p0_FIXED model}
\usage{
\method{generate}{TWEES_OCC_DAMPED_p0_FIXED}(x, new_data, specials = NULL, ...)
}
\arguments{
\item{x}{A fitted \code{TWEES_OCC_DAMPED_p0_FIXED} model object.}

\item{new_data}{A tsibble containing future index values to forecast.}

\item{specials}{Passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of future paths from a dataset using a fitted model.
}
\description{
Generate sample paths from a TWEES_OCC_DAMPED_p0_FIXED model
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)
fit <- model(ts, TWEES_OCC_DAMPED_p0_FIXED(value))
generate(fit, new_data = tsibble::new_data(ts, 7))
}
