% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marwal.R
\name{MARWAL}
\alias{MARWAL}
\title{Markov Chain Model with Random Walk dynamic}
\usage{
MARWAL(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
A Random Walk model on the demand composed by a swithcing constant and an ARMA model,
where the changes are controlled by a Markov Chain on the occurrence process.
Paramter are estimated in closed-form on deseasonalized data,
and forecasts are returned as Gaussian distributions for each time step.
}
\references{
Sbrana, G. (2025). Markov Walk and Walmart sales prediction.
\emph{Journal of the Operational Research Society}, 1–12.
}
