% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspes.R
\name{model_sum.HSPES}
\alias{model_sum.HSPES}
\title{Return model name}
\usage{
\method{model_sum}{HSPES}(x)
}
\arguments{
\item{x}{A fitted \code{fable} model object.}
}
\value{
The model name as a string
}
\description{
Return model name
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)
fit <- model(ts, HSPES(value))
model_sum(fit[[1]][[1]])
}
