% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_tweedie.R
\name{ptweedie}
\alias{ptweedie}
\title{Compute Tweedie CDF}
\usage{
ptweedie(
  q,
  mean = 1,
  dispersion = 1,
  power = 1.5,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{q}{The values the CDF is evaluated at.}

\item{mean}{Mean parameter of the distribution.}

\item{dispersion}{Dispersion parameter of the distribution.}

\item{power}{Power parameter of the distribution.}

\item{lower.tail}{Whether to return the lower tail probability.}

\item{log.p}{Whether to return the logarithm of the probability.}
}
\description{
Evaluates the Tweedie cumulative distribution function via the Compound Poisson-Gamma
representation, dispatching to the C++ implementation \code{tweedieCDF}.
}
\keyword{internal}
