% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_tweedie.R
\name{qtweedie}
\alias{qtweedie}
\title{Compute the Tweedie quantile function}
\usage{
qtweedie(
  p,
  mean = 1,
  dispersion = 1,
  power = 1.5,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{p}{The probabilities the quantile function is evaluated at.}

\item{mean}{Mean parameter of the distribution.}

\item{dispersion}{Dispersion parameter of the distribution.}

\item{power}{Power parameter of the distribution.}

\item{lower.tail}{Whether \code{p} represents lower tail probabilities.}

\item{log.p}{Whether \code{p} is given as log-probabilities.}
}
\description{
Evaluates the Tweedie quantile function via the Newton-Raphson method based on the CDF
and density function implementations, dispatching to the C++ implementation \code{tweedieInvCDF}.
}
\keyword{internal}
