% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile.dist_inflated}
\alias{quantile.dist_inflated}
\title{Quantile method for inflated distributions}
\usage{
\method{quantile}{dist_inflated}(x, p, ...)
}
\arguments{
\item{x}{An inflated distribution.}

\item{p}{A numeric vector of probabilities.}

\item{...}{Additional arguments passed to \code{\link[stats:quantile]{stats::quantile()}}.}
}
\value{
A numeric vector of quantiles.
}
\description{
Quantile method for inflated distributions
}
\examples{
x <- distributional::dist_inflated(distributional::dist_normal(), 0.2, 0)
stats::quantile(x, p = c(0.25, 0.5, 0.75))
}
\keyword{internal}
