% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raf}
\alias{raf}
\title{RAF Spare Parts Demand Dataset}
\format{
A tsibble with 420,000 rows and 3 variables:
\describe{
\item{series_id}{Character. Unique identifier for each time series.}
\item{index}{Date (yearmonth). The monthly time index.}
\item{value}{Numeric. The demand quantity for the given month.}
}
}
\source{
Syntetos, A. A., & Boylan, J. E. (2005). The accuracy of
intermittent demand estimates. \emph{International Journal of Forecasting},
21(2), 303--314.

Available at
\url{https://github.com/canerturkmen/gluon-ts/tree/intermittent-datasets/datasets}.
}
\usage{
raf
}
\description{
A dataset of monthly demand for spare parts from the Royal Air Force (RAF).
The data contains 5000 intermittent time series, each spanning 84 monthly
periods from January 1996 to December 2002. This is a widely used benchmark
dataset for intermittent demand forecasting.
}
\examples{
library(tsibble)
raf
}
\keyword{datasets}
