% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twees_occ_damped_p0_fixed.R
\name{residuals.TWEES_OCC_DAMPED_p0_FIXED}
\alias{residuals.TWEES_OCC_DAMPED_p0_FIXED}
\title{Extract residuals from a TWEES_OCC_DAMPED_p0_FIXED model}
\usage{
\method{residuals}{TWEES_OCC_DAMPED_p0_FIXED}(object, ...)
}
\arguments{
\item{object}{A model for which residuals are required.}

\item{...}{Not used.}
}
\value{
A numeric vector of residuals.
}
\description{
Extract residuals from a TWEES_OCC_DAMPED_p0_FIXED model
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)
fit <- model(ts, TWEES_OCC_DAMPED_p0_FIXED(value))
residuals(fit)
}
