% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_tweedie.R
\name{rtweedie}
\alias{rtweedie}
\title{Sample from Tweedie Distribution}
\usage{
rtweedie(n, mean = 1, dispersion = 1, power = 1.5)
}
\arguments{
\item{n}{The number of samples.}

\item{mean}{Mean parameter of the distribution.}

\item{dispersion}{Dispersion parameter of the distribution.}

\item{power}{Power parameter of the distribution.}
}
\description{
Uses the compound Poisson--Gamma representation to draw exact samples
for power in \eqn{(1, 2)} (Dunn & Smyth, 2005).
}
\keyword{internal}
