% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wss.R
\name{WSS}
\alias{WSS}
\title{Willemain--Smart--Schwarz Bootstrap Method}
\usage{
WSS(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
Markov-chain bootstrap forecasting method for intermittent demand proposed by
Willemain, Smart & Schwarz (2004). Occurrence of non-zero demand is modelled
with a two-state Markov chain whose transition probabilities are estimated
from the observed binary occurrence sequence. Demand sizes are resampled from
past non-zero observations with Gaussian jittering to smooth the empirical
distribution.
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)

fc_ts <- ts |>
  model(WSS(value)) |>
  forecast(h = "7 days")

fc_ts |> print()


if (requireNamespace("ggtime", quietly = TRUE)) {
  library(ggtime)
  fc_ts |> autoplot(ts)
}
}
\references{
Willemain, T. R., Smart, C. N., & Schwarz, H. F. (2004). A new approach to
forecasting intermittent demand for service parts inventories.
\emph{International Journal of Forecasting}, 20(3), 375--387.
\doi{10.1016/S0169-2070(03)00013-X}.
}
