% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betanbb.R
\name{BETANBB}
\alias{BETANBB}
\title{Beta-Negative Binomial Bayesian Dynamic Model}
\usage{
BETANBB(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
Conjugate Bayesian dynamic model for count time series with a negative
binomial observation distribution and a Beta prior on the success probability
parameter. This extends the conjugate updating framework of Harvey &
Fernandes (1989) to the Beta-Negative Binomial family. The Beta prior is
updated at each time step using a discount factor \code{w}. Forecasts are available
as samples simulating from the model forward in time.
}
\references{
Harvey, A. C., & Fernandes, C. (1989). Time series models for count or
qualitative observations. \emph{Journal of Business & Economic Statistics},
7(4), 407--417.
}
