% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.intermittent-package.R
\docType{package}
\name{fable.intermittent-package}
\alias{fable.intermittent}
\alias{fable.intermittent-package}
\title{fable.intermittent: Forecasting Models for Intermittent Time Series}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extends the 'fable' framework to support forecasting methods specifically designed for intermittent time series data, where demand occurs sporadically with many zero values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StefanoDamato/fable.intermittent}
  \item Report bugs at \url{https://github.com/StefanoDamato/fable.intermittent/issues}
}

}
\author{
\strong{Maintainer}: Stefano Damato \email{stefano.damato@supsi.ch} (\href{https://orcid.org/0009-0001-0225-3736}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Stefano Damato \email{stefano.damato@supsi.ch} (\href{https://orcid.org/0009-0001-0225-3736}{ORCID}) [copyright holder]
  \item Lorenzo Zambon \email{lorenzo.zambon@idsia.ch} (\href{https://orcid.org/0000-0002-8939-993X}{ORCID})
  \item Dario Azzimonti \email{dario.azzimonti@gmail.com} (\href{https://orcid.org/0000-0001-5080-3061}{ORCID})
}

}
\keyword{internal}
