% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.intermittent-package.R
\docType{package}
\name{fable.intermittent-package}
\alias{fable.intermittent}
\alias{fable.intermittent-package}
\title{fable.intermittent: Forecasting Models for Intermittent Time Series}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extends the 'fable' framework to support forecasting methods specifically designed for intermittent time series data, where demand occurs sporadically with many zero values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yourusername/fable.intermittent}
  \item Report bugs at \url{https://github.com/yourusername/fable.intermittent/issues}
}

}
\author{
\strong{Maintainer}: Stefano Damato \email{your.email@example.com}

}
\keyword{internal}
