% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gampoisb.R
\name{GAMPOISB}
\alias{GAMPOISB}
\alias{report.GAMPOISB}
\title{Gamma-Poisson Bayesian Dynamic Model}
\usage{
GAMPOISB(formula, ...)

\method{report}{GAMPOISB}(object, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}

\item{object}{A fitted model object.}
}
\value{
A model specification.
}
\description{
Conjugate Bayesian dynamic model for count time series with a Poisson
observation distribution and a Gamma prior on the rate parameter, following
Harvey & Fernandes (1989). The Gamma prior is updated at each time step
using a discount factor \code{w} that controls how quickly past information
decays. The first-step forecast follows a Negative Binomial distribution,
and multi-step forecasts are obtained by simulating from the model forward in time.
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)

fc_ts <- ts |>
  model(GAMPOISB(value)) |>
  forecast(h = "7 days")

fc_ts |> print()


if (requireNamespace("ggtime", quietly = TRUE)) {
  library(ggtime)
  fc_ts |> autoplot(ts)
}
}
\references{
Harvey, A. C., & Fernandes, C. (1989). Time series models for count or
qualitative observations. \emph{Journal of Business & Economic Statistics},
7(4), 407--417. \doi{10.1080/07350015.1989.10509750}.
}
