% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marwal.R
\name{MARWAL}
\alias{MARWAL}
\title{Markov Chain Model with Random Walk dynamic}
\usage{
MARWAL(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
A Random Walk model on the demand composed by a switching constant and an ARMA model,
where the changes are controlled by a Markov Chain on the occurrence process.
Parameters are estimated in closed-form on deseasonalized data,
and forecasts are returned as Gaussian distributions for each time step.
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)

fc_ts <- ts |>
  model(MARWAL(value)) |>
  forecast(h = "7 days")

fc_ts |> print()


if (requireNamespace("ggtime", quietly = TRUE)) {
  library(ggtime)
  fc_ts |> autoplot(ts)
}
}
\references{
Sbrana, G. (2025). Markov Walk and Walmart sales prediction.
\emph{Journal of the Operational Research Society}, 1--12. \doi{10.1080/01605682.2025.2569661}.
}
