% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnarma.R
\name{NNARMA}
\alias{NNARMA}
\alias{report.NNARMA}
\title{Non-negative ARMA model}
\usage{
NNARMA(formula, ...)

\method{report}{NNARMA}(object, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}

\item{object}{A fitted model object.}
}
\value{
A model specification.
}
\description{
#TODO
}
\examples{
ts <- tsibble::tsibble(
  time = as.Date("2026-01-01") + seq_len(40),
  value = rnbinom(40, size = 1, prob = 0.3),
  index = time
)

fc_ts <- ts |>
  model(NNARMA(value)) |>
  forecast(h = "7 days")

fc_ts |> print()


if (requireNamespace("ggtime", quietly = TRUE)) {
  library(ggtime)
  fc_ts |> autoplot(ts)
}
}
\references{
TODO
}
