% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twees.R
\name{TWEES}
\alias{TWEES}
\title{Tweedie Exponential Smoothing}
\usage{
TWEES(formula, damped = TRUE, scaling = TRUE, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{damped}{Logical. If \code{TRUE} (default), the exponential smoothing
component uses a damping parameter.}

\item{scaling}{Logical. If \code{TRUE} (default), the time series is divided by
its maximum value before fitting and predictions are back-transformed.
This improves numerical stability.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
Exponential smoothing state space model for intermittent demand with a
Tweedie observation distribution. The conditional mean of the Tweedie
is governed by a (optionally damped) exponential smoothing process.
The power and dispersion parameter are estimated to maximise the likelihood.
The Tweedie family naturally models both zeros and large spikes
via its compound Poisson-Gamma nature. The first-step forecast follows
a Tweedie distribution, and multi-step forecasts are obtained by simulating
from the model forward in time. The model parameters are estimated by
}
\references{
Damato, S., Azzimonti D., & Corani, G. (2025). Forecasting intermittent
time series with Gaussian Processes and Tweedie likelihood.
\emph{International Journal of Forecasting}  xx(x), xxx--xxx.
}
